+++ src/main/java/source/code/dto/response/workout/WorkoutResponseDto.java
@@ -7,7 +7,6 @@ import lombok.Setter;
 import source.code.dto.response.workoutSet.WorkoutSetResponseDto;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -15,10 +14,10 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class WorkoutResponseDto implements Serializable {
-    private int id;
+    private Integer id;
     private String name;
-    private BigDecimal duration;
-    private int restDaysAfter;
+    private Short duration;
+    private Byte restDaysAfter;
     private Integer weekIndex;
     private Integer dayOfWeekIndex;
     private List<WorkoutSetResponseDto> workoutSets;
